Grow decimation to uint32_t to fully support navg in the COR format.

+++ src/formats/drx.hpp
@@ -150,7 +150,7 @@ public:
         header->sync_word        = 0x5CDEC0DE;
         // ID is stored in the lowest 8 bits; bit 2 is reserved
         header->frame_count_word = htobe32((uint32_t) (hdr_base->src & 0xBF) << 24);
-        header->decimation       = htobe16(hdr_base->decimation);
+        header->decimation       = htobe16((uint16_t) hdr_base->decimation);
         header->time_offset      = 0;
         header->time_tag         = htobe64(hdr_base->seq);
         header->tuning_word      = htobe32(hdr_base->tuning);
